fix(users): limit 增加类型转换
(cherry picked from commit 08f1966b0f6441b6d1d00856ef3602340e3a6540)

+++ app/Api/Controller/Users/RecommendedUserController.php
@@ -76,14 +76,13 @@ class RecommendedUserController extends AbstractListController
     }
 
     /**
-     * 我的关注
      * {@inheritdoc}
      * @throws InvalidParameterException|InvalidArgumentException
      */
     protected function data(ServerRequestInterface $request, Document $document)
     {
         $include = $this->extractInclude($request);
-        $limit = Arr::get($request->getQueryParams(), 'limit');
+        $limit = (int)Arr::get($request->getQueryParams(), 'limit');
 
         $cacheKey = 'users_recommendedUser';
         $cacheData = $this->cache->get($cacheKey);
